+++ plone/namedfile/tests/__init__.py
@@ -1,0 +1,7 @@
+import os
+
+
+def getFile(filename):
+    """ return contents of the file with the given name """
+    filename = os.path.join(os.path.dirname(__file__), filename)
+    return open(filename, 'rb')