+++ plone/namedfile/tests/test_scaling_functional.py
@@ -8,22 +8,16 @@ from plone.namedfile.file import NamedImage
 from plone.namedfile.interfaces import IImageScaleTraversable
 from plone.namedfile.scaling import ImageScaling
 from plone.namedfile.testing import PLONE_NAMEDFILE_FUNCTIONAL_TESTING
+from plone.namedfile.tests import getFile
 from plone.testing.z2 import Browser
 from StringIO import StringIO
 from zope.annotation import IAttributeAnnotatable
 from zope.interface import implementer
 
-import os
 import PIL
 import time
 import transaction
 import unittest
-
-
-def getFile(filename):
-    """ return contents of the file with the given name """
-    filename = os.path.join(os.path.dirname(__file__), filename)
-    return open(filename, 'rb')
 
 
 def wait_to_ensure_modified():